refactor: Complete migration from internal/tui to internal/ui
Move TUI components to more organized UI structure:
- internal/tui/chat/ → internal/ui/terminal/
- internal/tui/handlers/ → internal/ui/handlers/
- internal/tui/render/ → internal/ui/render/
- internal/tui/styles/ → internal/ui/styles/

Update all import paths and package names to reflect new structure.
Update documentation to match new organization.
