+++ internal/ui/render/chat.go
