+++ internal/ui/terminal/handlers.go
@@ -1,4 +1,4 @@
-package chat
+package terminal
 
 import (
 	"context"
