+++ internal/ui/terminal/types.go
@@ -1,4 +1,4 @@
-package chat
+package terminal
 
 import "github.com/mizzy/rigel/internal/llm"
